fixes bugs

+++ routes.js
@@ -8,14 +8,14 @@ const bcrypt = require('bcrypt')
 
 /* Handler function to wrap each route. */
 function asyncHandler(cb){
-    return async(req, res, next) => {
-      try {
-        await cb(req, res, next)
-      } catch(error){
-        // Forward error to the global error handler
-        next(error);
-      }
+  return async(req, res, next) => {
+    try {
+      await cb(req, res, next)
+    } catch(error){
+      // Forward error to the global error handler
+      next(error);
     }
+  }
 }
 
 // Construct a router instance.
@@ -35,14 +35,19 @@ router.get('/users', authenticateUser, asyncHandler(async (req, res) => {
 // POST Route that creates a new user
 router.post('/users', asyncHandler(async (req, res) => {
     try {
-        const hashedPassword = bcrypt.hashSync(req.body.password, 10);
-        await User.create({
+        const user = await User.create({
           firstName: req.body.firstName,
           lastName: req.body.lastName,
           emailAddress: req.body.emailAddress,
-          password: hashedPassword
+          password: req.body.password
         });
-        res.status(201).json({ "message": "Account successfully created!" });
+
+        await user.update({
+          password: bcrypt.hashSync(user.password, 10)
+        })
+
+        res.location('/').status(201);
+        res.end();
     } catch (error) {
         if (error.name === 'SequelizeValidationError' || error.name === 'SequelizeUniqueConstraintError') {
           const errors = error.errors.map(err => err.message);
@@ -56,7 +61,11 @@ router.post('/users', asyncHandler(async (req, res) => {
 // GET Route that returns all courses available
 router.get('/courses', asyncHandler(async (req, res) => {
     let courses = await Course.findAll({
-      attributes: ['id', 'title', 'description', 'estimatedTime', 'materialsNeeded']
+      attributes: ['id', 'title', 'description', 'estimatedTime', 'materialsNeeded'],
+      include: [{
+        model: User,
+        attributes: ['id', 'firstName', 'lastName', 'emailAddress']
+      }]  
     })
     res.status(200).json(courses)
 }));
@@ -65,9 +74,13 @@ router.get('/courses', asyncHandler(async (req, res) => {
 router.get('/courses/:id', asyncHandler(async (req, res) => {
   let course = await Course.findAll({
     where: {
-        id: req.params.id
+        id: req.params.id,
     },
     attributes: ['id', 'title', 'description', 'estimatedTime', 'materialsNeeded'],
+    include: [{
+      model: User,
+      attributes: ['id', 'firstName', 'lastName', 'emailAddress']
+    }]
   });
   res.status(200).json(course)
 }));
@@ -75,8 +88,10 @@ router.get('/courses/:id', asyncHandler(async (req, res) => {
 // POST Route that creates a course
 router.post('/courses', authenticateUser, asyncHandler(async (req, res) => {
   try {
-      await Course.create(req.body);
-      res.status(201).json({ "message": "Course successfully created!" });
+      const course = await Course.create(req.body);
+      console.log(course.id)
+      res.location(`/courses/${course.id}`).status(201);
+      res.end()
   } catch (error) {
       if (error.name === 'SequelizeValidationError' || error.name === 'SequelizeUniqueConstraintError') {
         const errors = error.errors.map(err => err.message);
@@ -89,43 +104,55 @@ router.post('/courses', authenticateUser, asyncHandler(async (req, res) => {
 
 // PUT Route that will update the corresponding course
 router.put('/courses/:id', authenticateUser, asyncHandler(async (req, res) => {
-  let course = await Course.findAll({
+  const course = await Course.findAll({
     where: {
         id: req.params.id
     }
   })
-  // && course.title && course.description
-  if(course && req.body.title && req.body.description){
-    await Course.update({
-      title: req.body.title,
-      description: req.body.description,
-      estimatedTime: req.body.estimatedTime,
-      materialsNeeded: req.body.materialsNeeded,
-      userId: req.body.userId
-    }, {
-      where: {id: req.params.id}
-    })
-    res.status(204).end();
-  } else {
-    if(course){
-      res.status(400).json({message: "Title nor Descriptioin can be empty"});
-    }else{
-      res.status(404).json({message: "Course Not Found"});
+
+  // first checks if the course belongs to the current user, then if the course exist and title/description is filled in.
+  if(req.currentUser.id === course[0].dataValues.id){
+    if(course && req.body.title && req.body.description){
+      await Course.update({
+        title: req.body.title,
+        description: req.body.description,
+        estimatedTime: req.body.estimatedTime,
+        materialsNeeded: req.body.materialsNeeded,
+        userId: req.body.userId
+      }, {
+        where: {id: req.params.id}
+      })
+      res.status(204).end();
+    } else {
+      let errors = new Error(); 
+      if(course){
+        errors.message = "Title nor Descriptioin can be empty"
+        res.status(400).json({ errors });
+      }else{
+        res.status(404).json({message: "Course Not Found"});
+      }
     }
+  } else {
+    res.status(403).json({ message: 'Access Denied' });
   }
 }));
 
 // DELETE Route that will delete the corresponding course
 router.delete('/courses/:id', authenticateUser, asyncHandler(async (req, res) => {
   let course = await Course.findAll({
-    where: {id: req.params.id}
+    where: {id: req.params.id},
   });
 
-  if(course){
-    Course.destroy({ where: {id: req.params.id} });
-    res.status(204).end();
+  // first checks if course belongs to the current user, then if course exists:
+  if(req.currentUser.id === course[0].dataValues.id){
+    if(course){
+      Course.destroy({ where: {id: req.params.id} });
+      res.status(204).end();
+    } else {
+      res.status(404).json({message: "Course Not Found"});
+    }
   } else {
-    res.status(404).json({message: "Course Not Found"});
+    res.status(403).json({ message: 'Access Denied' });
   }
 }));
